Deconstructed preview into separate components to help create build page
