+++ src/components/section/preivew/index.js
@@ -4,12 +4,20 @@ import BuildField from './BuildField'
 import BuildItem from './BuildItem'
 import BuildList from './BuildList'
 import BuildSubmitBtn from './BuildSubmitBtn'
+import VerifyResponse from './VerifyResponse'
+import DecodeBodyObject from './DecodeBodyObject'
+import DisplayAlert from './DisplayAlert'
+import BuildFieldArea from './BuildFieldArea'
 
 export {
     BuildField,
+    BuildFieldArea,
     BuildItem,
     BuildList,
     BuildSubmitBtn,
     Constant,
-    SectionPreview
+    DecodeBodyObject,
+    DisplayAlert,
+    SectionPreview,
+    VerifyResponse
 }
